refactor: home 컴포넌트 모듈화 작업으로인한 수정
