+++ src/routes/Router.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { Layout } from '../components/layout';
 import ApplicationPage from '../pages/application/Application';
-import { Home } from '../pages/home';
+import Home from '../pages/home/Home';
 
 const Router: React.FC = () => (
 	<BrowserRouter>
